+++ kitchen.global.yml
@@ -17,8 +17,8 @@ verifier:
 platforms:
   - name: almalinux-8
   - name: almalinux-9
+  - name: amazonlinux-2023
   - name: centos-stream-9
-  - name: debian-12
   - name: debian-13
   - name: fedora-latest
   - name: oraclelinux-8
