+++ kitchen.yml
@@ -3,27 +3,24 @@ driver:
   name: vagrant
 
 provisioner:
-  name: chef_infra
+  name: chef_zero
   deprecations_as_errors: true
   chef_license: accept
   product_name: chef
   product_version: <%= ENV['CHEF_VERSION'] || 'current' %>
-  install_strategy: once
-  enforce_idempotency: <%= ENV['ENFORCE_IDEMPOTENCY'] || true %>
-  multiple_converge: <%= ENV['MULTIPLE_CONVERGE'] || 2 %>
+  install_strategy: always
   log_level: <%= ENV['CHEF_LOG_LEVEL'] || 'auto' %>
 
 verifier:
   name: inspec
-  sudo: true
 
 platforms:
   - name: almalinux-8
   - name: almalinux-9
-  - name: fedora-latest
+  - name: amazonlinux-2023
   - name: centos-stream-9
-  - name: debian-12
   - name: debian-13
+  - name: fedora-latest
   - name: oraclelinux-8
   - name: oraclelinux-9
   - name: rockylinux-8
@@ -31,10 +28,17 @@ platforms:
   - name: ubuntu-22.04
   - name: ubuntu-24.04
 
+x-run_lists:
+  default: &default_run_list
+    - recipe[test::default]
+
+x-verifiers:
+  default: &default_verifier
+    inspec_tests:
+      - path: test/integration/default
+
 suites:
   - name: default
-    run_list:
-      - recipe[test::default]
-    verifier:
-      inspec_tests:
-        - path: test/integration/default
+    run_list: *default_run_list
+    verifier: *default_verifier
+...
