+++ libraries/helpers.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 #
 # Cookbook:: unbound
 # Library:: helpers
@@ -18,6 +19,20 @@
 module Unbound
   module Cookbook
     module Helpers
+      def deepsort?
+        return if defined?(DeepSort)
+
+        begin
+          Gem::Specification.find_by_name('deepsort')
+        rescue Gem::MissingSpecError
+          declare_resource(:chef_gem, 'deepsort')
+        end
+
+        require 'deepsort'
+
+        true
+      end
+
       def default_config_dir
         return '/etc/unbound' if %i(unbound_config unbound_configure unbound_config_server).include?(declared_type)
 
@@ -52,6 +67,37 @@ module Unbound
           'no'
         when 'yes', 'YES', 'no', 'NO'
           value.downcase
+        end
+      end
+
+      def perform_config_action(config)
+        if %i(create create_if_missing).include?(new_resource.action)
+          directory new_resource.config_dir do
+            owner new_resource.owner
+            group new_resource.group
+            mode new_resource.directory_mode
+            recursive true
+            action :create
+          end
+        end
+
+        config.merge!(new_resource.extra_options.dup) unless new_resource.extra_options.empty?
+
+        if new_resource.sort
+          deepsort?
+          config.deep_sort!
+        end
+
+        template new_resource.config_file do
+          cookbook new_resource.cookbook
+          source new_resource.template
+          owner new_resource.owner
+          group new_resource.group
+          mode new_resource.mode
+          sensitive new_resource.sensitive
+          helpers(Unbound::Cookbook::TemplateHelpers)
+          variables(content: config)
+          action new_resource.action
         end
       end
     end
