+++ metadata.rb
@@ -13,8 +13,9 @@ chef_version     '>= 16'
 gem 'deepsort'
 
 supports 'almalinux', '>= 8.0'
+supports 'amazon', '>= 2023.0'
 supports 'centos_stream', '>= 9.0'
-supports 'debian', '>= 12.0'
+supports 'debian', '>= 13.0'
 supports 'fedora'
 supports 'oracle', '>= 8.0'
 supports 'redhat', '>= 8.0'
