+++ resources/config_authority_zone.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 #
 # Cookbook:: unbound
 # Resource:: config_authority_zone
@@ -15,14 +16,13 @@
 # limitations under the License.
 #
 
-# frozen_string_literal: true
-
-unified_mode true
-
 provides :unbound_config_authority_zone
 provides :unbound_config_auth_zone
+unified_mode true
 
-use 'partials/_config_file'
+include Unbound::Cookbook::Helpers
+
+use '_partial/_config_file'
 
 property :config_file, String,
           default: lazy { "#{config_dir}/authority-zone-#{name}.conf" },
@@ -72,6 +72,8 @@ load_current_value do |new_resource|
 end
 
 action_class do
+  include Unbound::Cookbook::Helpers
+
   def do_template_action
     zone_config = {
       'name' => new_resource.zone_name,
@@ -94,3 +96,5 @@ action_class do
     perform_config_action(config)
   end
 end
+
+%i(create create_if_missing delete).each { |action_type| action(action_type) { do_template_action } }
