+++ resources/config_dnstap.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 #
 # Cookbook:: unbound
 # Resource:: config_dnstap
@@ -15,13 +16,12 @@
 # limitations under the License.
 #
 
-# frozen_string_literal: true
-
+provides :unbound_config_dnstap
 unified_mode true
 
-provides :unbound_config_dnstap
+include Unbound::Cookbook::Helpers
 
-use 'partials/_config_file'
+use '_partial/_config_file'
 
 property :config_file, String,
           default: lazy { "#{config_dir}/dnstap.conf" },
@@ -88,6 +88,8 @@ load_current_value do |new_resource|
 end
 
 action_class do
+  include Unbound::Cookbook::Helpers
+
   def do_template_action
     zone_config = {
       'dnstap-enable' => new_resource.dnstap_enable,
@@ -118,3 +120,5 @@ action_class do
     perform_config_action(config)
   end
 end
+
+%i(create create_if_missing delete).each { |action_type| action(action_type) { do_template_action } }
