+++ test/cookbooks/test/recipes/default.rb
@@ -2,22 +2,14 @@
 
 unbound_package 'unbound'
 
-server_config = {
-  verbosity: 1,
-  interface: '127.0.0.1',
-}
-
-if platform_family?('rhel', 'fedora')
-  server_config.merge!(
-    'chroot' => '',
-    'username' => 'unbound',
-    'directory' => '/etc/unbound',
-    'pidfile' => '/var/run/unbound/unbound.pid'
-  )
-end
-
 unbound_config 'unbound' do
-  server server_config
+  server({
+           verbosity: 1,
+           interface: [
+             '127.0.0.1',
+             '127.0.0.1@853',
+           ],
+         })
   notifies :restart, 'unbound_service[unbound]', :delayed
 end
 
@@ -35,5 +27,10 @@ unbound_config_remote_control 'unbound-remote-control' do
 end
 
 unbound_service 'unbound' do
-  action [:enable, :start]
+  action :enable
+end
+
+unbound_service 'unbound start' do
+  service_name 'unbound'
+  action :start
 end
