+++ test/integration/default/inspec.yml
@@ -1,7 +1,7 @@
 ---
 name: default
-title: Unbound Default Tests
+title: Default Tests
 maintainer: Sous Chefs
 license: Apache-2.0
-summary: Verify Unbound package, configuration, and service resources
+summary: Default integration tests for the unbound cookbook
 version: 1.0.0
